renamed package for LEGO model
