+++ core/src/main/java/com/github/jlgrock/snp/core/classifier/AssertionClassifier.java
@@ -3,7 +3,7 @@ package com.github.jlgrock.snp.core.classifier;
 import org.jvnet.hk2.annotations.Contract;
 
 import com.github.jlgrock.snp.core.domain.ClassifiedAssertion;
-import com.github.jlgrock.snp.core.model.parser.Assertion;
+import com.github.jlgrock.snp.core.model.xml.lego.Assertion;
 
 /**
  * Classify an entire assertion object
