--- core/src/main/java/com/github/jlgrock/snp/core/model/parser/package-info.java
@@ -1,4 +1,0 @@
-/**
- *
- */
-package com.github.jlgrock.snp.core.model.parser;
