+++ core/src/main/java/com/github/jlgrock/snp/core/model/xml/lego/Lego.java
@@ -1,4 +1,4 @@
-package com.github.jlgrock.snp.core.model.parser;
+package com.github.jlgrock.snp.core.model.xml.lego;
 
 import com.google.common.base.MoreObjects;
