+++ core/src/main/java/com/github/jlgrock/snp/core/model/xml/lego/package-info.java
@@ -1,0 +1,4 @@
+/**
+ *
+ */
+package com.github.jlgrock.snp.core.model.xml.lego;
