+++ web/src/test/java/com/github/jlgrock/snp/web/controllers/AssertionClassifierServiceImplTest.java
@@ -14,7 +14,7 @@ import org.testng.annotations.Test;
 import com.github.jlgrock.snp.core.classifier.AssertionClassifier;
 import com.github.jlgrock.snp.core.data.ClassifiedAssertionStore;
 import com.github.jlgrock.snp.core.domain.ClassifiedAssertion;
-import com.github.jlgrock.snp.core.model.parser.Assertion;
+import com.github.jlgrock.snp.core.model.xml.lego.Assertion;
 
 public class AssertionClassifierServiceImplTest {
